Task done 100%
We extended the Cart.js controller, with the help of AllAi we manage to see what is in the cart and retrieve the total price. From there we created a custom site preference, so we can easily update the threshold without changing the code. Then we created a content asset that holds the message that appear in the page, add some CSS and it is ready