Done 100%
Now when you click on the first image you get the first card element and the same for the second. Had some problem with the images selection, but change the images classes to be different. You can change the image from the business manager, because it is content asset, and the two images are in content slot, thats how we render them